fix: add dummy dependencies to prevent concurrent execution (#611)
* fix: change project to lava-dev-436720 in .df-credentials.json

Otherwise `make dataform-run` fails

* fix: add dummy dependencies to prevent concurrent execution

There seems to be a difference in the way that the dataform cli command
and the API handle bigquery limits. This is is leading to a limits exceeded
error when we run dataform in ci but not during dev.
Other people have encountered this error
(fromt two weeks ago: https://www.googlecloudcommunity.com/gc/Data-Analytics/Dataform-Exceeded-rate-limits-too-many-updates-for-this-routine/m-p/804841#M7384)
but there is no fix yet.

This PR simply adds some unecessary dependencies between sqlx files
so that they run one-at-the-time.  This should keep us from hitting the
rate limits.  In the future, when dataform API is working like the cli
command, we can remove these dependencies.