refactor: change 'expire' to 'mature' in facility (#1494)
In the context of a facility, 'expired' usually refers to the date that
no more drawdowns are allowed from, whereas 'matured' referes to the
date where outstanding principal payments are due.

In our domain, these are the same thing, but the more relevant concept
for our context is the payment due date, so 'mature' makes more sense as
the word to use.

This refactor would also have the benefit or re-introducing the concept
of 'expired' if we decided to stop drawdowns at some point before the
facility matures.