fix(reporting): filter stale rows from raw tables (#1958)
This is a hack to fix an issue with reports where rows from
prior rollouts persist in tables and create records that should
not be there and do not match the state in the runtime.

The hack works by recognizing that the chart of accounts is only initialized once
so we can use it to know when the database was reset.